fix: extract BatchedBroadcaster to own file to avoid web-assets CI dependency
The test imported from server.ts which imports web-assets.generated — a
file that only exists after building web assets. CI runs tests before
building, causing a module resolution failure.
